==> scripts/02-data-preprocessing-python.ipynb <==
!pip install pyreadr

import pandas as pd
import numpy as np
from pyreadr import read_r
import re
from string import ascii_lowercase

# Load the lres.RDS file
result = read_r("../raw-data/lres.RDS")

result.head()

print(result.keys())
